Split Qt MainWindow into its own file
